decouple core module from carriers for better admin models section naming
