+++ purpleserver/core/purpleserver/carriers/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CarriersConfig(AppConfig):
+    name = 'carriers'
