+++ purpleserver/core/purpleserver/carriers/models.py
@@ -1,12 +1,34 @@
 import pkgutil
 import logging
-import purpleserver.core.extension.models as extensions
+from functools import partial
 from typing import Any, Dict
+
 from django.db import models
-from purpleserver.core.models.carrier import Carrier
 from purplship.package import gateway
+from purpleserver.core.models import Entity, uuid
+import purpleserver.carriers.extension.models as extensions
 
 logger = logging.getLogger(__name__)
+
+
+class Carrier(Entity):
+    id = models.CharField(max_length=50, primary_key=True, default=partial(uuid, prefix='car_'), editable=False)
+    carrier_id = models.CharField(
+        max_length=200, unique=True,
+        help_text="eg. canadapost, dhl_express, fedex, purolator_courrier, ups..."
+    )
+    test = models.BooleanField(default=True)
+
+    def __str__(self):
+        return self.carrier_id
+
+    def settings(self):
+        for field in [f for f in self._meta.get_fields() if isinstance(f, models.OneToOneRel)]:
+            try:
+                return getattr(self, field.get_accessor_name())
+            except:
+                pass
+        return None
 
 
 class CanadaPostSettings(Carrier):
