+++ purpleserver/core/purpleserver/carriers/router.py
@@ -1,0 +1,3 @@
+from rest_framework.routers import DefaultRouter
+
+router = DefaultRouter(trailing_slash=False)
