+++ purpleserver/core/purpleserver/carriers/urls.py
@@ -1,0 +1,10 @@
+"""
+purplship server carriers module urls
+"""
+from django.urls import include, path
+from purpleserver.carriers.views import router
+
+app_name = 'purpleserver.carriers'
+urlpatterns = [
+    path('v1/', include(router.urls)),
+]
