+++ purpleserver/core/purpleserver/core/admin.py
@@ -1,5 +1,1 @@
 from django.contrib import admin
-import purpleserver.core.models as models
-
-for name, model in models.MODELS.items():
-    admin.site.register(model)
