--- purpleserver/core/purpleserver/core/models/__init__.py
@@ -1,1 +1,0 @@
-from purpleserver.core.models.settings import *
