+++ purpleserver/core/purpleserver/core/serializers.py
@@ -1,4 +1,4 @@
-from purpleserver.core.models import MODELS
+from purpleserver.carriers.models import MODELS
 from purplship.core.units import (
     Country, WeightUnit, DimensionUnit, PackagingUnit, PaymentType, Currency, PrinterType
 )
