+++ purpleserver/core/purpleserver/core/views/__init__.py
@@ -1,4 +1,3 @@
-import purpleserver.core.views.carriers
-import purpleserver.core.views.utils
 import purpleserver.core.views.labels
+import purpleserver.core.views.references
 from purpleserver.core.router import router
