+++ purpleserver/purpleserver/settings.py
@@ -52,6 +52,7 @@ static_lazy = lazy(static, str)
 
 PURPLSHIP_CONF = [app for app in [
     {'app': 'purpleserver.core', 'urls': 'purpleserver.core.urls'},
+    {'app': 'purpleserver.carriers', 'urls': 'purpleserver.carriers.urls'},
     {'app': 'purpleserver.proxy', 'urls': 'purpleserver.proxy.urls'},
     {'app': 'purpleserver.manager', 'urls': 'purpleserver.manager.urls'},
 ] if importlib.util.find_spec(app['app']) is not None]
